+++ src/main/java/com/example/Pitlance/Models/SellerModelAndDTO/SellerNameEmailBalancePhonePasswordTPI.java
@@ -2,7 +2,7 @@ package com.example.Pitlance.Models.SellerModelAndDTO;
 
 import java.math.BigDecimal;
 
-public record SellerNameEmailBalancePhonePasswordTPI(String companyName, String sellerName, String sellerLastName, String email, int phoneNumber, String taxPayerId, BigDecimal balance, String password) {
+public record SellerNameEmailBalancePhonePasswordTPI(String companyName, String sellerName, String sellerLastName, String email, long phoneNumber, String taxPayerId, BigDecimal balance, String password) {
     public static SellerNameEmailBalancePhonePasswordTPI of(Seller seller){
         return new SellerNameEmailBalancePhonePasswordTPI(seller.getCompanyName(), seller.getSellerName(), seller.getSellerLastName(), seller.getEmail(), seller.getPhoneNumber(), seller.getTaxPayerId(), seller.getBalance(), seller.getPassword());
     }
